Move answer generation to knowpro/answers.py; move some utilities to aitools/utils.py
