+++ python/ta/typeagent/aitools/auth.py
@@ -58,21 +58,6 @@ def get_shared_token_provider() -> AzureTokenProvider:
     return _shared_token_provider
 
 
-def load_dotenv() -> None:
-    dirname = os.path.dirname
-    here = dirname(__file__)
-    typeagent = dirname(here)
-    package = dirname(typeagent)
-    python = dirname(package)
-    repo_top = dirname(python)
-    env_path = os.path.join(repo_top, "ts", ".env")
-    dotenv.load_dotenv(env_path)
-    # for k, v in os.environ.items():
-    #     if "KEY" in k:
-    #         print(f"{k}={v!r}")
-    # print(f"Loaded {env_path}")
-
-
 if __name__ == "__main__":
     # Usage: eval `./typeagent/aitools/auth.py`
     print(f"export AZURE_OPENAI_API_KEY={AzureTokenProvider().get_token()}")
